refactor: userInfo를 불러오는 과정에서의 중복코드 줄이기 (#933)
* refactor: requestGetUserInfo가 user와 auth에서 중복적으로 존재하여 auth에 존재하는 함수 삭제

* refactor: useRequestSuspenseGetUserInfo 삭제 및 useUserInfoContext를 활용하여 userInfo 불러오는 것으로 변경

* chore: useRequestGetUserInfo을 hooks/queries/auth에서 user로 위치 변경

---------
